requirements.txt
# Core Django
Django==5.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
django-filter==23.5

# Database
psycopg2-binary==2.9.9

# Authentication & Security
django-otp==1.3.0
qrcode[pil]==7.4.2

# File handling & Media
Pillow==10.2.0
django-cleanup==8.0.0

# Reports & Export
reportlab==4.0.9
openpyxl==3.1.2
django-import-export==3.3.7

# API Documentation
drf-spectacular==0.27.1

# Background tasks
celery==5.3.6
redis==5.0.1
django-celery-beat==2.5.0

# Email & Notifications
django-anymail==10.2

# Environment & Configuration
django-environ==0.11.2
python-decouple==3.8

# Utilities
python-dateutil==2.8.2
pytz==2024.1

# Production Server
gunicorn==21.2.0
whitenoise==6.6.0

# Development & Testing
pytest==7.4.4
pytest-django==4.7.0
factory-boy==3.3.0
faker==22.6.0
